Refactor function call
