+++ src/sema/expression/constructor.rs
@@ -549,3 +549,34 @@ pub fn new(
         init: None,
     })
 }
+
+/// Is it an (new C).value(1).gas(2)(1, 2, 3) style constructor (not supported)?
+pub(super) fn deprecated_constructor_arguments(
+    expr: &pt::Expression,
+    diagnostics: &mut Diagnostics,
+) -> Result<(), ()> {
+    match expr.remove_parenthesis() {
+        pt::Expression::FunctionCall(func_loc, ty, _) => {
+            if let pt::Expression::MemberAccess(_, ty, call_arg) = ty.as_ref() {
+                if deprecated_constructor_arguments(ty, diagnostics).is_err() {
+                    // location should be the identifier and the arguments
+                    let mut loc = call_arg.loc;
+                    if let pt::Loc::File(_, _, end) = &mut loc {
+                        *end = func_loc.end();
+                    }
+                    diagnostics.push(Diagnostic::error(
+                        loc,
+                        format!("deprecated call argument syntax '.{}(...)' is not supported, use '{{{}: ...}}' instead", call_arg.name, call_arg.name)
+                    ));
+                    return Err(());
+                }
+            }
+        }
+        pt::Expression::New(..) => {
+            return Err(());
+        }
+        _ => (),
+    }
+
+    Ok(())
+}
